src/main/python/nwselect/requirements.txt
requests == 2.26.0
pyyaml